refactor: simplify Rigs Without Cigs add information page with server-side validation
